requirements.txt
# Core
pytz==2018.9
Django==2.2.28
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.12.0.113

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1
django-picklefield==2.1.1
django-treebeard==4.3
django-json-widget==0.2.0
django-admin-sortable2==0.7.6

# Rest apis
djangorestframework==3.11.2
djangorestframework-csv==2.1.0
django-rest-auth==0.9.5
Markdown==3.0.1
django-filter==2.1.0
django-cors-headers==3.2.0

# Developer Tools
ipdb==0.11
ipython==7.31.1
flake8==3.7.5
django-debug-toolbar==2.2.1
sentry-sdk==0.16.1
django-extensions==3.1.3
Werkzeug==2.0.2

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.2
coverage==4.5.2
pytest==5.4.1
pytest-django==3.9.0
django-mock-queries==2.1.5
django-test-plus==1.4.0
selenium==4.1.3

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.93
Pillow==9.0.1 # Needed for ImageField

# Serving static content
whitenoise==5.0.1
django-cors-headers==3.2.0

# GIS libraries
GDAL==3.0.2
djangorestframework-gis==0.14
Fiona==1.8.13.post1
Shapely==1.6.4

django-q==1.3.6
redis==3.5.3

# Data manipulation
pandas==1.0.0
xlrd==1.2.0
django-import-export==2.0.2
django-map-widgets==0.3.0
django-material-icon-widget==0.1.2
django-colorfield==0.3.2

# amazon-s3
django-storages==1.7.1
boto3==1.9.93

# object-permissions
django-guardian==2.2.0

django-mock-queries==2.1.5
lorem==0.1.1
chardet==3.0.4

django-admin-json-editor==0.2.3
django-tinymce==3.4.0
django-simple-history==3.0.0
